RPN-728: on backend support signatures with v = 31/32, to support hardware wallets Trezor/Ledger which do not implement EIP-712.
Trezor/Ledger do not have support for EIP-712, thus they cannot call 'eth_signTypedData_v4' or similar. Thus, we need a workaround: encode and hash the 'Order' struct manually according to the EIP-712 and then sign it using the wallet's private key via 'eth_sign' method. The 'eth_sign' method adds the '\x19Ethereum Signed Message:' prefix to the hash, and we need to handle that in the signature validator on backend.